Reubicacion Vistas incorrectas, ahora son componentes
