+++ src/components/formularios/index.ts
@@ -3,3 +3,4 @@ export { default as VinculadosForm } from "./VinculadosForm.vue";
 export { default as EfectosForm } from "./EfectosForm.vue";
 export { default as PersonalIntervinienteComponent } from "./PersonalIntervinienteComponent.vue";
 export { default as FechaUbicacionComponent } from "./FechaUbicacionComponent.vue";
+export { default as FormActuacionComponent } from "./FormActuacionComponent.vue";
