--- src/views/AfectadosView.vue
@@ -1,8 +1,0 @@
-<script setup lang="ts">
-import { AfectadosForm } from '@/components/formularios/index';
-</script>
-<template>
-  <div class="w-full">
-    <AfectadosForm />
-  </div>
-</template>
