--- src/views/EfectosView.vue
@@ -1,8 +1,0 @@
-<script lang="ts" setup>
-import { EfectosForm } from '@/components/formularios/index';
-</script>
-<template>
-  <div class="w-full">
-    <EfectosForm />
-  </div>
-</template>
