--- src/views/FechaUbicacionView.vue
@@ -1,7 +1,0 @@
-<script setup lang="ts">
-import { FechaUbicacionComponent } from '@/components/formularios/index'
-</script>
-
-<template>
-  <FechaUbicacionComponent />
-</template>
