+++ src/views/FormActuacionView.vue
@@ -1,0 +1,6 @@
+<script setup lang="ts">
+import { FormActuacionComponent } from '@/components/formularios/index'
+</script>
+<template>
+  <FormActuacionComponent />
+</template>
