--- src/views/PersonalInterviniente.vue
@@ -1,7 +1,0 @@
-<script lang="ts" setup>
-import { PersonalIntervinienteComponent } from '@/components/formularios/index'
-</script>
-
-<template>
-  <PersonalIntervinienteComponent />
-</template>
