--- src/views/VinculadosView.vue
@@ -1,8 +1,0 @@
-<script setup lang="ts">
-import { VinculadosForm } from '@/components/formularios/index';
-</script>
-<template>
-  <div>
-    <VinculadosForm />
-  </div>
-</template>
