+++ src/views/index.ts
@@ -1,0 +1,5 @@
+export { default as AccessDeniedView } from "./AccessDeniedView.vue";
+export { default as ActuacionesView } from "./ActuacionesView.vue";
+export { default as ActuacionView } from "./ActuacionView.vue";
+export { default as DatosLegalesView } from "./DatosLegalesView.vue";
+export { default as FormActuacionView } from "./FormActuacionView.vue";
